Change Vehicle Id property type from int to long in migration snapshots; update README with migration instructions

+++ Migrations/BoxedSoftware.Postgres/Migrations/VehiclesContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace BoxedSoftware.Postgres.Migrations
 
             modelBuilder.Entity("BoxedSoftware.Models.Vehicle", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Fuel")
                         .IsRequired()
